requirements.txt
# Core dependencies
PyGithub>=2.1.1
python-dotenv>=1.0.0

# RAG-specific dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.0.2
langchain-community>=0.0.13
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Text processing
tiktoken>=0.5.2